專題後端/requirements.txt
Flask
flask-cors
joblib
scikit-learn
requests
python-dotenv
